requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
# Tkinter is built-in with Python, no additional installation needed